Add utils to API calls and save data
